Separate IRequestLoader file
