+++ packages/core/src/storages/sitemap_request_loader.ts
@@ -13,7 +13,7 @@ import type { CrawleeLogger } from '../log.js';
 import { Request } from '../request.js';
 import { serviceLocator } from '../service_locator.js';
 import { KeyValueStore } from './key_value_store.js';
-import type { IRequestLoader } from './request_list.js';
+import type { IRequestLoader } from './request_loader.js';
 import type { IRequestManager } from './request_provider.js';
 import { purgeDefaultStorages } from './utils.js';
